+++ ja2lib/Strategic/CreatureSpreading.h
@@ -27,7 +27,7 @@ extern BOOLEAN gfUseCreatureMusic;
 BOOLEAN MineClearOfMonsters(UINT8 ubMineIndex);
 
 // Returns TRUE if valid and creature quest over, FALSE if creature quest active or not yet started
-BOOLEAN GetWarpOutOfMineCodes(INT16 *psSectorX, INT16 *psSectorY, INT8 *pbSectorZ,
+BOOLEAN GetWarpOutOfMineCodes(u8 *psSectorX, u8 *psSectorY, INT8 *pbSectorZ,
                               INT16 *psInsertionGridNo);
 
 extern INT16 gsCreatureInsertionCode;
